re did seed files and migration to get rid of total available column and current total column on event_users, also built out all the CRUD functionaliyt for events, and event users

+++ models/event.js
@@ -127,7 +127,11 @@ class Event {
       }
   }
   destroyEvent(event_id) {
-    return database.raw(`DELETE FROM events WHERE id = ?`, event_id)
+    if (this.validateEventExists(event_id) === false) {
+      return false
+    } else {
+      return database.raw(`DELETE FROM events WHERE id = ?`, event_id)
+    }
   }
 }
